if_ath: ath_tgt_send_beacon assert if beacon exist
With current code we will print error and try to execute something
If bhdr will be accidantly not NULL, then we will continue with some
trash without letting host knwo about it.

In this case it is better to use assert and send to the host at least panic
notification.

Signed-off-by: Oleksij Rempel <linux@rempel-privat.de>