fix compile error with gcc-10
With GCC 10.2 we will get following error:
toolchain/inst/lib/gcc/xtensa-elf/10.2.0/../../../../xtensa-elf/bin/ld:
  target_firmware/build/k2/libfirmware.a(init.c.obj):target_firmware/magpie_fw_dev/target/init/init.h:53:
  multiple definition of `_assfail_ori';
  target_firmware/build/k2/libfirmware.a(app_start.c.obj):target_firmware/magpie_fw_dev/target/init/init.h:53:
  first defined here

Since previous compilers merged this variable automatically, the GCC 10
will do it only if we will tell it to do so. With this patch we should
be able to compile this firmware with new and old compiler.

Signed-off-by: Oleksij Rempel <linux@rempel-privat.de>